enable awesome map block
