+++ app/cells/decidim/decidim_awesome/content_blocks/map/error.erb
@@ -1,0 +1,11 @@
+<section class="wrapper-home map-block home-section">
+  <div class="row">
+    <div class="columns">
+      <div class="callout alert">
+        <p><%= t("unavailable", scope: "decidim.decidim_awesome.map_component.map.error") %></p>
+      </div>
+
+      <%= render partial: "decidim/decidim_awesome/map_component/map/api_ready.html" %>
+    </div>
+  </div>
+</div>
