+++ app/controllers/decidim/decidim_awesome/map_component/map_controller.rb
@@ -5,21 +5,13 @@ module Decidim
     module MapComponent
       class MapController < DecidimAwesome::BlankComponentController
         helper Decidim::DecidimAwesome::MapHelper
-        helper_method :map_components, :api_ready?
+        helper_method :map_components
 
         def show
-          render :error unless maps_enabled?
+          render :error unless Decidim::Map.configured?
         end
 
         private
-
-        def maps_enabled?
-          Decidim::Map.configured?
-        end
-
-        def api_ready?
-          Decidim::Api::Schema.max_complexity >= 1300
-        end
 
         def map_components
           @map_components ||= current_participatory_space.components.published.filter do |component|
