+++ app/views/decidim/decidim_awesome/map_component/map/_api_ready.html.erb
